+++ workspace/solutions/kotlin/lib/src/test/kotlin/s0/s0/s4/s3/s_0043_test.kt
@@ -1,0 +1,16 @@
+package s0.s0.s4.s3
+
+import utils.Test_
+import kotlin.test.Test
+import s0.s0.s4.s3.s_0043_00.Solution as s_0043_00
+
+class s_0043_test {
+    @Test
+    fun test() {
+        val test = Test_(
+            this::class,
+            s_0043_00::class,
+        )
+        test.run()
+    }
+}