+++ workspace/solutions/rust/src/s0/s0/s4/s3/mod.rs
@@ -1,0 +1,4 @@
+pub mod s_0043_00;
+pub mod s_0043_test;
+
+pub type TSolution = fn(String, String) -> String;