+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - 添加题目 `0415` 的题解 | Add the solution to the question `0415`
   - [415. 字符串相加 - 力扣（LeetCode）](https://leetcode.cn/problems/add-strings/description/)
+- 添加题目 `0043` 的题解 | Add the solution to the question `0043`
+  - [43. 字符串相乘 - 力扣（LeetCode）](https://leetcode.cn/problems/multiply-strings/description/)
 
 ## 2024-08-23